Fork ruler-of-exchange for use in gmail
It will now output an xml-file that can be imported from the gmail web
interface.

I would prefer to use the google APIs, but they require an API key. This script
would not be very user friendly if every user needs to create an API key. I'll
therefore just accept the toil of having to upload the xml file every time I
make a change (for now).

Also changing the license to GPL-3, which (according to my admittedly limited
legal understanding) is compatible with the original do-what-the-fuck-you-want-to
license.